// src/lib/types.ts
export interface IUsers {
    id: number | string
    name: string
    surname: string
    age: number 
    salary: number 
}

export interface IForm {
    name: string
    surname: string
    age: number 
    salary: number
}

export interface IUser {
    id: number | string
    name: string
    surname: string
    age: string
    salary: string 
}